+++ src/components/Typography/Paragraph/types/Paragraph.type.ts
@@ -1,4 +1,0 @@
-import { CommonHTMLProps, MouseEvents } from '../../../../types';
-
-export type ParagraphProps = {} & CommonHTMLProps<HTMLParagraphElement> &
-    MouseEvents<HTMLParagraphElement>;